requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
langchain==0.1.12
langchain-community==0.0.27
langchain-google-genai==0.0.11
google-generativeai==0.3.2
python-dotenv==1.0.1
PyPDF2==3.0.1
chromadb==0.4.24 